+++ R/write_src.R
@@ -3,9 +3,10 @@ write_src <- function(path) {
   src_path <- file.path(path, "src")
   dir.create(src_path)
   write_utils(src_path)
-  write_rmd(src_path, "Introduction", "01-introduction.Rmd")
-  write_rmd(src_path, "Methods", "02-methods.Rmd")
-  write_rmd(src_path, "Results", "03-results.Rmd")
+  write_setup(src_path)
+  write_rmd(src_path, "Introduction", 1)
+  write_rmd(src_path, "Methods", 2)
+  write_rmd(src_path, "Results", 3)
 }
 
 # utils
@@ -22,7 +23,19 @@ write_utils <- function(path) {
              con = file.path(path, "utils.R"))
 }
 
+# setup
+write_setup <- function(path) {
+  writeLines(text = "# Setup script", con = file.path(path, "setup.R"))
+}
+
 # Rmd
-write_rmd <- function(path, header, rmd_name) {
-  writeLines(text = paste("#", header), con = file.path(path, rmd_name))
+write_rmd <- function(path, name, number) {
+  setup_chunk <- ymlthis::code_chunk(
+    chunk_code = source(here::here("src/setup.R"), encoding = "UTF-8"),
+    chunk_name = paste0("setup-", sprintf("%02d", number)),
+    chunk_args = list(include = FALSE)
+  )
+  rmd_name <- paste0(sprintf("%02d", number), "-", tolower(name), ".Rmd")
+  writeLines(text = paste0("# ", name, "\n\n", setup_chunk),
+             con = file.path(path, rmd_name))
 }